+++ changelog.md
@@ -2,6 +2,7 @@ TGUI 1.9  (TBD)
 ---------------
 
 - New backend: SDL_GPU (uses SDL3's new GPU API)
+- RichTextLabel can now contain url tags and a new findLinkAtPos function can check if the mouse is on top of any
 - Added mapPixelToCoords and mapCoordsToPixel to CanvasSFML
 - Added tgui::literals::percent namespace with the "_percent" user-defined literal
 - Fixed behavior of using showWithEffect or hideWithEffect with a Fade type while another fade animation wasn't finished yet